+++ packages/gatsby-source-wordpress-experimental/.test-runtime/__tests__/integration/schema/data-resolution.test.js
@@ -1,234 +1,13 @@
 import fetchGraphql from "gatsby-source-wordpress-experimental/utils/fetch-graphql"
-import { introspectionQuery } from "gatsby-source-wordpress-experimental/utils/graphql-queries"
-
-import { runGatsby } from "../test-utils/run-gatsby"
-import gql from "gatsby-source-wordpress-experimental/utils/gql"
-import { incrementalIt } from "../test-utils/incremental-it"
-import { createContentDigest } from "gatsby-core-utils"
+import { incrementalIt } from "../../../test-utils/incremental-it"
 
 jest.setTimeout(100000)
 
-require(`dotenv`).config({
-  path: `./test-runtime/.env.development`,
-})
-
-require(`dotenv`).config({
-  path: `./test-runtime/credentials.env`,
-})
-
-describe(`[gatsby-source-wordpress-experimental] schema`, () => {
-  runGatsby()
-
-  it(`hasn't altered the remote WPGraphQL schema`, async () => {
-    const result = await fetchGraphql({
-      query: introspectionQuery,
-      url: process.env.WPGRAPHQL_URL,
-    })
-
-    const remoteWPGQLTypeNamesWithFieldNames = result.data.__schema.types.map(
-      type => ({
-        name: type.name,
-        fields:
-          type && type.fields ? type.fields.map(field => field.name) : null,
-      })
-    )
-
-    expect(remoteWPGQLTypeNamesWithFieldNames).toMatchSnapshot()
-
-    expect(createContentDigest(result.data.__schema)).toMatchSnapshot()
-  })
-
-  it(`hasn't altered the local Gatsby schema`, async () => {
-    const result = await fetchGraphql({
-      url: `http://localhost:8000/___graphql`,
-      query: introspectionQuery,
-    })
-
-    const localWPTypeNamesWithFieldNames = result.data.__schema.types
-      .filter(({ name }) => name.startsWith(`allWp`) || name.startsWith(`Wp`))
-      .map(type => ({
-        name: type.name,
-        fields:
-          type && type.fields ? type.fields.map(field => field.name) : null,
-      }))
-
-    expect(localWPTypeNamesWithFieldNames).toMatchSnapshot()
-
-    expect(createContentDigest(result.data.__schema)).toMatchSnapshot()
-  })
-
-  test(`Type.exclude option removes types from the schema`, async () => {
-    const result = await fetchGraphql({
-      url: `http://localhost:8000/___graphql`,
-      query: gql`
-        {
-          __schema {
-            types {
-              name
-            }
-          }
-        }
-      `,
-    })
-
-    const excludedTypes = [
-      `ActionMonitorAction`,
-      `UserToActionMonitorActionConnection`,
-      `Plugin`,
-      `PostFormat`,
-      `Theme`,
-      `UserRole`,
-      `UserToUserRoleConnection`,
-    ]
-
-    // make sure our schema doesn't have any of the default excluded types
-    // in ../../src/models/gatsby-api
-    excludedTypes.forEach(typeName => {
-      expect(
-        !!result.data.__schema.types.find(type => type.name === `Wp${typeName}`)
-      ).toBe(false)
-    })
-  })
-
-  test(`Type.excludeFieldNames removes fields from the schema on types by field names`, async () => {
-    const excludeFieldsOnTypesGroups = [
-      {
-        typeName: `Page`,
-        fieldNames: [`enclosure`],
-      },
-      {
-        typeName: `User`,
-        fieldNames: [`extraCapabilities`, `capKey`, `email`, `registeredDate`],
-      },
-    ]
-
-    const query = `
-        {
-          ${excludeFieldsOnTypesGroups
-            .map(
-              group => `
-                ${group.typeName}: __type(name: "Wp${group.typeName}") {
-                  fields {
-                    name
-                  }
-                }
-              `
-            )
-            .join(` `)}
-        }
-      `
-
-    const result = await fetchGraphql({
-      url: `http://localhost:8000/___graphql`,
-      query,
-    })
-
-    // for all our test type names
-    excludeFieldsOnTypesGroups.forEach(excludeGroup => {
-      // get that type from the Gatsby schema
-      const typeInSchema = result.data[excludeGroup.typeName]
-
-      // for each excluded field on our type
-      excludeGroup.fieldNames.forEach(fieldName => {
-        // expect that the type in our Gatsby schema doesn't have a field
-        // with this excluded name
-        const fieldNameExistsOnType = !!typeInSchema.fields
-          .map(field => field.name)
-          .includes(fieldName)
-
-        expect(fieldNameExistsOnType).toBe(false)
-      })
-    })
-  })
-
-  test(`excludeFieldNames option excludes fields from the schema globally`, async () => {
-    const result = await fetchGraphql({
-      url: `http://localhost:8000/___graphql`,
-      query: gql`
-        {
-          wpPage: __type(name: "WpPage") {
-            fields {
-              name
-            }
-          }
-
-          wpPost: __type(name: "WpPost") {
-            fields {
-              name
-            }
-          }
-        }
-      `,
-    })
-
-    const commentFieldFinder = ({ name }) => name === `commentCount`
-
-    const wpPageCommentCountField = result.data.wpPage.fields.find(
-      commentFieldFinder
-    )
-
-    const wpPostCommentCountField = result.data.wpPost.fields.find(
-      commentFieldFinder
-    )
-
-    expect(wpPostCommentCountField && wpPageCommentCountField).toBeFalsy()
-  })
-
-  test(`Type.where option works when set to filter for French posts`, async () => {
-    const result = await fetchGraphql({
-      url: `http://localhost:8000/___graphql`,
-      query: gql`
-        {
-          allWpTranslationFilterTest {
-            totalCount
-            nodes {
-              title
-            }
-          }
-        }
-      `,
-    })
-
-    expect(result.data.allWpTranslationFilterTest.totalCount).toEqual(1)
-    expect(result.data.allWpTranslationFilterTest.nodes[0].title).toEqual(
-      `French page`
-    )
-  })
-
-  test(`Type.limit option works when set to 1`, async () => {
-    const result = await fetchGraphql({
-      url: `http://localhost:8000/___graphql`,
-      query: gql`
-        {
-          allWpTypeLimitTest {
-            totalCount
-          }
-        }
-      `,
-    })
-
-    expect(result.data.allWpTypeLimitTest.totalCount).toEqual(1)
-  })
-
-  test(`Type.limit option works when set to 0`, async () => {
-    const result = await fetchGraphql({
-      url: `http://localhost:8000/___graphql`,
-      query: gql`
-        {
-          allWpTypeLimit0Test {
-            totalCount
-          }
-        }
-      `,
-    })
-
-    expect(result.data.allWpTypeLimit0Test.totalCount).toEqual(0)
-  })
-
+describe(`[gatsby-source-wordpress-experimental] data resolution`, () => {
   incrementalIt(`resolves menus`, async () => {
     const result = await fetchGraphql({
       url: `http://localhost:8000/___graphql`,
-      query: gql`
+      query: /* GraphQL */ `
         {
           allWpMenu {
             nodes {
@@ -282,7 +61,7 @@ describe(`[gatsby-source-wordpress-experimental] schema`, () => {
   incrementalIt(`resolves pages`, async () => {
     const result = await fetchGraphql({
       url: `http://localhost:8000/___graphql`,
-      query: gql`
+      query: /* GraphQL */ `
         {
           testPage: wpPage(id: { eq: "cGFnZToy" }) {
             title
@@ -400,7 +179,7 @@ describe(`[gatsby-source-wordpress-experimental] schema`, () => {
   incrementalIt(`resolves posts`, async () => {
     const result = await fetchGraphql({
       url: `http://localhost:8000/___graphql`,
-      query: gql`
+      query: /* GraphQL */ `
         {
           testPost: wpPost(id: { eq: "cG9zdDox" }) {
             title
@@ -438,7 +217,7 @@ describe(`[gatsby-source-wordpress-experimental] schema`, () => {
   incrementalIt(`resolves users`, async () => {
     const result = await fetchGraphql({
       url: `http://localhost:8000/___graphql`,
-      query: gql`
+      query: /* GraphQL */ `
         {
           testUser: wpUser(id: { eq: "dXNlcjox" }) {
             firstName
@@ -473,7 +252,7 @@ describe(`[gatsby-source-wordpress-experimental] schema`, () => {
   incrementalIt(`resolves root fields`, async () => {
     const result = await fetchGraphql({
       url: `http://localhost:8000/___graphql`,
-      query: gql`
+      query: /* GraphQL */ `
         {
           wp {
             allSettings {