+++ packages/gatsby-source-wordpress-experimental/.test-runtime/test-utils/global-teardown-jest.js
@@ -1,0 +1,19 @@
+// global-teardown.js
+import kill from "tree-kill"
+import { resetSchema } from "./increment-remote-data"
+
+module.exports = async function globalTeardown() {
+  if (!process.env.START_SERVER) {
+    return
+  }
+
+  const gatsbyProcess = global.__GATSBY_PROCESS
+
+  kill(gatsbyProcess.pid)
+  console.log(`\nkilled Gatsby`)
+
+  if (process.env.WPGQL_INCREMENT) {
+    await resetSchema()
+    console.log(`reset remote api mutations`)
+  }
+}